New commit
NunchuckCharacterization is used for evaluating the nunchuck limits and calibrating the code to fit these.
ServoPulseCalibrator is used to calibrate the servo limits and discover the upper and lower values to store these for use in the integration code.
EyesIntegration_Enhanced_v4 is the final version of the code with servo safeguard, startup animation, regular use and idle animation.